+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Teknoo Software - Recipe - Change Log
 
+
+## [3.3.0] - 2021-08-12
+### Stable Release
+- Add `Promise::fetchResult`, To get the returned value by the callback on the promise (Can be null if the callback 
+  return nothing). *(Not east compliant, but useful to integrate east code with an non-east code).*
+
 ## [3.2.1] - 2021-08-19
 ### Stable Release
 - QA, Fix PHPDoc.